Service and controllers for Food and Recipe repositories

+++ src/main/java/com/foodrestapi/recipes/repository/FoodRepository.java
@@ -3,5 +3,5 @@ package com.foodrestapi.recipes.repository;
 import com.foodrestapi.recipes.model.Food;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoodRepository extends JpaRepository<Food, Integer> {
+public interface FoodRepository extends JpaRepository<Food, Long> {
 }
